INLauncherb-0.1/INLauncher-main/requirements.txt
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
attrs==25.4.0
beautifulsoup4==4.14.2
binaryornot==0.4.4
bs4==0.0.2
certifi==2025.6.15
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
cryptography==46.0.3
fastapi==0.115.13
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
flet-web==0.28.3
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
minecraft-launcher-lib==7.1
multidict==6.7.0
NBT==1.5.1
oauthlib==3.3.1
packaging==25.0
pefile==2023.2.7
pillow==12.0.0
plyer==2.1.0
propcache==0.4.1
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyGithub==2.8.1
Pygments==2.19.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.10.1
PyNaCl==1.6.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
Python-NBT==1.2.0
python-slugify==8.0.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.46.2
text-unidecode==1.3
toml==0.10.2
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
watchdog==4.0.2
watchfiles==1.1.0
websockets==15.0.1
yarl==1.22.0
